Edited existing entities to be integrated with login system, clean architecture refactoring
