+++ src/main/java/app/Main.java
@@ -12,7 +12,6 @@ public class Main {
                 .addSignupUseCase()
                 .addLoginUseCase()
                 .addLogoutUseCase()
-                .addChangePasswordUseCase()
                 .build();
 
         application.pack();
