+++ src/main/java/data_access/FileUserDataAccessObject.java
@@ -1,8 +1,7 @@
 package data_access;
 
-import entity.User;
-import entity.UserFactory;
-import use_case.change_password.ChangePasswordUserDataAccessInterface;
+import entity.Household;
+import entity.HouseholdFactory;
 import use_case.login.LoginUserDataAccessInterface;
 import use_case.logout.LogoutUserDataAccessInterface;
 import use_case.signup.SignupUserDataAccessInterface;
@@ -17,24 +16,23 @@ import java.util.Map;
  */
 public class FileUserDataAccessObject implements SignupUserDataAccessInterface,
                                                  LoginUserDataAccessInterface,
-                                                 ChangePasswordUserDataAccessInterface,
                                                  LogoutUserDataAccessInterface {
 
     private static final String HEADER = "username,password";
 
     private final File csvFile;
     private final Map<String, Integer> headers = new LinkedHashMap<>();
-    private final Map<String, User> accounts = new HashMap<>();
+    private final Map<String, Household> accounts = new HashMap<>();
 
     private String currentUsername;
 
     /**
      * Construct this DAO for saving to and reading from a local file.
      * @param csvPath the path of the file to save to
-     * @param userFactory factory for creating user objects
+     * @param householdFactory factory for creating user objects
      * @throws RuntimeException if there is an IOException when accessing the file
      */
-    public FileUserDataAccessObject(String csvPath, UserFactory userFactory) {
+    public FileUserDataAccessObject(String csvPath, HouseholdFactory householdFactory) {
 
         csvFile = new File(csvPath);
         headers.put("username", 0);
@@ -57,8 +55,8 @@ public class FileUserDataAccessObject implements SignupUserDataAccessInterface,
                     final String[] col = row.split(",");
                     final String username = String.valueOf(col[headers.get("username")]);
                     final String password = String.valueOf(col[headers.get("password")]);
-                    final User user = userFactory.create(username, password);
-                    accounts.put(username, user);
+                    final Household household = householdFactory.create(username, password);
+                    accounts.put(username, household);
                 }
             }
             catch (IOException ex) {
@@ -74,9 +72,9 @@ public class FileUserDataAccessObject implements SignupUserDataAccessInterface,
             writer.write(String.join(",", headers.keySet()));
             writer.newLine();
 
-            for (User user : accounts.values()) {
+            for (Household household : accounts.values()) {
                 final String line = String.format("%s,%s",
-                        user.getName(), user.getPassword());
+                        household.getHouseholdID(), household.getPassword());
                 writer.write(line);
                 writer.newLine();
             }
@@ -90,13 +88,13 @@ public class FileUserDataAccessObject implements SignupUserDataAccessInterface,
     }
 
     @Override
-    public void save(User user) {
-        accounts.put(user.getName(), user);
+    public void save(Household household) {
+        accounts.put(household.getHouseholdID(), household);
         this.save();
     }
 
     @Override
-    public User get(String username) {
+    public Household get(String username) {
         return accounts.get(username);
     }
 
@@ -115,10 +113,4 @@ public class FileUserDataAccessObject implements SignupUserDataAccessInterface,
         return accounts.containsKey(identifier);
     }
 
-    @Override
-    public void changePassword(User user) {
-        // Replace the User object in the map
-        accounts.put(user.getName(), user);
-        save();
-    }
 }
