+++ src/main/java/data_access/InMemoryUserDataAccessObject.java
@@ -1,7 +1,6 @@
 package data_access;
 
-import entity.User;
-import use_case.change_password.ChangePasswordUserDataAccessInterface;
+import entity.Household;
 import use_case.login.LoginUserDataAccessInterface;
 import use_case.logout.LogoutUserDataAccessInterface;
 import use_case.signup.SignupUserDataAccessInterface;
@@ -15,10 +14,9 @@ import java.util.Map;
  */
 public class InMemoryUserDataAccessObject implements SignupUserDataAccessInterface,
                                                      LoginUserDataAccessInterface,
-                                                     ChangePasswordUserDataAccessInterface,
-                                                     LogoutUserDataAccessInterface {
+                                                    LogoutUserDataAccessInterface {
 
-    private final Map<String, User> users = new HashMap<>();
+    private final Map<String, Household> users = new HashMap<>();
 
     private String currentUsername;
 
@@ -28,12 +26,12 @@ public class InMemoryUserDataAccessObject implements SignupUserDataAccessInterfa
     }
 
     @Override
-    public void save(User user) {
-        users.put(user.getName(), user);
+    public void save(Household household) {
+        users.put(household.getHouseholdID(), household);
     }
 
     @Override
-    public User get(String username) {
+    public Household get(String username) {
         return users.get(username);
     }
 
@@ -45,12 +43,6 @@ public class InMemoryUserDataAccessObject implements SignupUserDataAccessInterfa
     @Override
     public String getCurrentUsername() {
         return currentUsername;
-    }
-
-    @Override
-    public void changePassword(User user) {
-        // Replace the old entry with the new password
-        users.put(user.getName(), user);
     }
 
 }
