+++ src/main/java/entity/HouseholdFactory.java
@@ -1,0 +1,11 @@
+package entity;
+
+/**
+ * Factory for creating CommonUser objects.
+ */
+public class HouseholdFactory {
+
+    public Household create(String name, String password) {
+        return new Household(name, password);
+    }
+}
