--- src/main/java/entity/UserFactory.java
@@ -1,11 +1,0 @@
-package entity;
-
-/**
- * Factory for creating CommonUser objects.
- */
-public class UserFactory {
-
-    public User create(String name, String password) {
-        return new User(name, password);
-    }
-}
