--- src/main/java/interface_adapter/logged_in/ChangePasswordPresenter.java
@@ -1,33 +1,0 @@
-package interface_adapter.logged_in;
-
-import interface_adapter.ViewManagerModel;
-import use_case.change_password.ChangePasswordOutputBoundary;
-import use_case.change_password.ChangePasswordOutputData;
-
-/**
- * The Presenter for the Change Password Use Case.
- */
-public class ChangePasswordPresenter implements ChangePasswordOutputBoundary {
-
-    private final LoggedInViewModel loggedInViewModel;
-    private final ViewManagerModel viewManagerModel;
-
-    public ChangePasswordPresenter(ViewManagerModel viewManagerModel,
-                             LoggedInViewModel loggedInViewModel) {
-        this.viewManagerModel = viewManagerModel;
-        this.loggedInViewModel = loggedInViewModel;
-    }
-
-    @Override
-    public void prepareSuccessView(ChangePasswordOutputData outputData) {
-        loggedInViewModel.getState().setPassword("");
-        loggedInViewModel.getState().setPasswordError(null);
-        loggedInViewModel.firePropertyChange("password");
-    }
-
-    @Override
-    public void prepareFailView(String error) {
-        loggedInViewModel.getState().setPasswordError(error);
-        loggedInViewModel.firePropertyChange("password");
-    }
-}
