--- src/main/java/use_case/change_password/ChangePasswordInteractor.java
@@ -1,37 +1,0 @@
-package use_case.change_password;
-
-import entity.User;
-import entity.UserFactory;
-
-/**
- * The Change Password Interactor.
- */
-public class ChangePasswordInteractor implements ChangePasswordInputBoundary {
-    private final ChangePasswordUserDataAccessInterface userDataAccessObject;
-    private final ChangePasswordOutputBoundary userPresenter;
-    private final UserFactory userFactory;
-
-    public ChangePasswordInteractor(ChangePasswordUserDataAccessInterface changePasswordDataAccessInterface,
-                                    ChangePasswordOutputBoundary changePasswordOutputBoundary,
-                                    UserFactory userFactory) {
-        this.userDataAccessObject = changePasswordDataAccessInterface;
-        this.userPresenter = changePasswordOutputBoundary;
-        this.userFactory = userFactory;
-    }
-
-    @Override
-    public void execute(ChangePasswordInputData changePasswordInputData) {
-        if ("".equals(changePasswordInputData.getPassword())) {
-            userPresenter.prepareFailView("New password cannot be empty");
-        }
-        else {
-            final User user = userFactory.create(changePasswordInputData.getUsername(),
-                    changePasswordInputData.getPassword());
-
-            userDataAccessObject.changePassword(user);
-
-            final ChangePasswordOutputData changePasswordOutputData = new ChangePasswordOutputData(user.getName());
-            userPresenter.prepareSuccessView(changePasswordOutputData);
-        }
-    }
-}
