+++ src/main/java/use_case/signup/SignupInteractor.java
@@ -1,7 +1,7 @@
 package use_case.signup;
 
-import entity.User;
-import entity.UserFactory;
+import entity.Household;
+import entity.HouseholdFactory;
 
 /**
  * The Signup Interactor.
@@ -9,14 +9,14 @@ import entity.UserFactory;
 public class SignupInteractor implements SignupInputBoundary {
     private final SignupUserDataAccessInterface userDataAccessObject;
     private final SignupOutputBoundary userPresenter;
-    private final UserFactory userFactory;
+    private final HouseholdFactory householdFactory;
 
     public SignupInteractor(SignupUserDataAccessInterface signupDataAccessInterface,
                             SignupOutputBoundary signupOutputBoundary,
-                            UserFactory userFactory) {
+                            HouseholdFactory householdFactory) {
         this.userDataAccessObject = signupDataAccessInterface;
         this.userPresenter = signupOutputBoundary;
-        this.userFactory = userFactory;
+        this.householdFactory = householdFactory;
     }
 
     @Override
@@ -34,10 +34,10 @@ public class SignupInteractor implements SignupInputBoundary {
             userPresenter.prepareFailView("Username cannot be empty");
         }
         else {
-            final User user = userFactory.create(signupInputData.getUsername(), signupInputData.getPassword());
-            userDataAccessObject.save(user);
+            final Household household = householdFactory.create(signupInputData.getUsername(), signupInputData.getPassword());
+            userDataAccessObject.save(household);
 
-            final SignupOutputData signupOutputData = new SignupOutputData(user.getName());
+            final SignupOutputData signupOutputData = new SignupOutputData(household.getHouseholdID());
             userPresenter.prepareSuccessView(signupOutputData);
         }
     }
