+++ src/main/java/use_case/signup/SignupUserDataAccessInterface.java
@@ -1,6 +1,6 @@
 package use_case.signup;
 
-import entity.User;
+import entity.Household;
 
 /**
  * DAO interface for the Signup Use Case.
@@ -16,7 +16,7 @@ public interface SignupUserDataAccessInterface {
 
     /**
      * Saves the user.
-     * @param user the user to save
+     * @param household the user to save
      */
-    void save(User user);
+    void save(Household household);
 }
